fix: 매니지먼트 모듈 리팩토링중
- common 모듈 수정 (version 체크 로직 추가)
- 응답 dto 및 요청 dto 교차 체크 (안드로이드) 후 수정

+++ core/src/main/java/com/mongs/core/vo/mqtt/PublishAttendanceVo.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record PublishAttendanceVo(
         Long mongId,
-        Integer exp
+        Double exp
 ) {
 }
